Преобразование сервиса корзины - разбиение на сервис корзины и сервис хранения данных корзины
