+++ Services/WebStore.Services/Product/CookiesCartStore.cs
@@ -1,0 +1,51 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Microsoft.AspNetCore.Http;
+using Newtonsoft.Json;
+using WebStore.Domain.Models;
+using WebStore.Interfaces.Services;
+
+namespace WebStore.Services.Product
+{
+    public class CookiesCartStore : ICartStore
+    {
+        private readonly string _CartName;
+        private readonly IHttpContextAccessor _HttpContextAccessor;
+
+        public Cart Cart
+        {
+            get
+            {
+                var context = _HttpContextAccessor.HttpContext;
+                var cookies = context.Response.Cookies;
+                var cart_cookie = context.Request.Cookies[_CartName];
+                if (cart_cookie is null)
+                {
+                    var cart = new Cart();
+                    cookies.Append(_CartName, JsonConvert.SerializeObject(cart));
+                    return cart;
+                }
+
+                ReplaceCookie(cookies, cart_cookie);
+                return JsonConvert.DeserializeObject<Cart>(cart_cookie);
+            }
+            set => ReplaceCookie(_HttpContextAccessor.HttpContext.Response.Cookies, JsonConvert.SerializeObject(value));
+        }
+
+        private void ReplaceCookie(IResponseCookies cookies, string cookie)
+        {
+            cookies.Delete(_CartName);
+            cookies.Append(_CartName, cookie, new CookieOptions { Expires = DateTime.Now.AddDays(15) });
+        }
+
+        public CookiesCartStore(IHttpContextAccessor HttpContextAccessor)
+        {
+            _HttpContextAccessor = HttpContextAccessor;
+
+            var user = HttpContextAccessor.HttpContext.User;
+            var user_name = user.Identity.IsAuthenticated ? user.Identity.Name : null;
+            _CartName = $"cart[{user_name}]";
+        }
+    }
+}
